Revamp leaderboards output
leaderboards now outputs a user-friendly embed. Not tested for multiple users yet, but it will still be better than what we have currently.
The command is still unfinished as the reactions don't do anything yet, and I don't know how to change the SQL to have a different number of rows.